[Doc Comment Translation] Improve Remarks Formatting
The focus of these changes is to improve formatting for the `<remarks>`
element of the XML doc comments to treat them as text rather than
pre-formatted blocks.

Previously, transformation was sensitive to the formatting of the source
code, attempting to preserve line breaks and spacing rather than treating
it as XML content.  The source code was often formatted to wrap lines so
that they could be more easily read by other developers, causing the
rendering for the MS Docs site to contain awkward and undesirable formatting.

With these changes, the `<remarks>` are treated as content, normalizing
line breaks and spacing and introducing explicit paragraphs to ensure that
the structure is semantically accurate and assist in HTML rendering.

This should address the ask in DevOps [#136807](https://ceapex.visualstudio.com/Engineering/_workitems/edit/136807)
as well as the choppy lines rendered for other Azure SDK types, such as the
[`EventDataBatch`](https://docs.microsoft.com/dotnet/api/azure.messaging.eventhubs.producer.eventdatabatch.tryadd?view=azure-dotnet#remarks)
class.