[Doc Comment Translation] Improve Remarks Formatting (#190)
* [Doc Comment Translation] Improve Remarks Formatting

The focus of these changes is to improve formatting for the `<remarks>`
element of the XML doc comments to treat them as text rather than
pre-formatted blocks.

Previously, transformation was sensitive to the formatting of the source
code, attempting to preserve line breaks and spacing rather than treating
it as XML content.  The source code was often formatted to wrap lines so
that they could be more easily read by other developers, causing the
rendering for the MS Docs site to contain awkward and undesirable formatting.

With these changes, the `<remarks>` are treated as content, normalizing
line breaks and spacing and introducing explicit paragraphs to ensure that
the structure is semantically accurate and assist in HTML rendering.

This should address the ask in DevOps [#136807](https://ceapex.visualstudio.com/Engineering/_workitems/edit/136807)
as well as the choppy lines rendered for other Azure SDK types, such as the
[`EventDataBatch`](https://docs.microsoft.com/dotnet/api/azure.messaging.eventhubs.producer.eventdatabatch.tryadd?view=azure-dotnet#remarks)
class.

* Ignore <remarks> wrapped in <format>

This set of changes updates processing for remarks to make use of the
existing logic in NormalizeDocsElement to detect the pattern:

<remarks>
  <format>
    ...
  </format>
</remarks>

When this pattern is detected, the remarks are considered preformatted
and parsing into paragraphs will be skipped.

* Don't Wrap Single Line In Paragraph Tags

Adjust the approach to not wrap text that is a single line without
embedded paragraphs with <p></p> tags; the MS Docs site will render
a single unwrapped line as a paragraph by default.